Jul 9 19:03, KvRequest -> lock -> txn request
