+++ Cargo.toml
@@ -33,6 +33,10 @@ futures-timer = "3.0"
 async-trait = "0.1.27"
 tokio = { version = "0.2", features = ["sync"] }
 
+tikv-client-common = {path = "tikv-client-common"}
+tikv-client-request = {path = "tikv-client-request"}
+tikv-client-pd = {path = "tikv-client-pd"}
+
 [dependencies.prometheus]
 version = "0.8"
 default-features = false
@@ -48,3 +52,10 @@ fail = { version = "0.3", features = [ "failpoints" ] }
 
 [patch.crates-io]
 raft-proto = { git = "https://github.com/tikv/raft-rs", rev = "e624c1d48460940a40d8aa69b5329460d9af87dd" }
+
+[workspace]
+members = [
+    "tikv-client-common",
+    "tikv-client-request",
+    "tikv-client-pd"
+]
