+++ tikv-client-common/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "tikv-client-common"
+version = "0.0.0"
+edition = "2018"
+
+
+[dependencies]
+failure = "0.1"
+grpcio = { version = "0.5.0", features = [ "secure", "prost-codec" ], default-features = false }
+futures = { version = "0.3.1", features = ["compat", "async-await", "thread-pool"] }
+kvproto = { git = "https://github.com/pingcap/kvproto.git", features = [ "prost-codec" ], default-features = false }
+lazy_static = "1"
+
+[dependencies.prometheus]
+version = "0.8"
+default-features = false
+features = ["push", "process"]
