+++ tikv-client-pd/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "tikv-client-pd"
+version = "0.0.0"
+edition = "2018"
+
+
+[dependencies]
+async-trait = "0.1.27"
+derive-new = "0.5"
+kvproto = { git = "https://github.com/pingcap/kvproto.git", features = [ "prost-codec" ], default-features = false }
