+++ tikv-client-request/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "tikv-client-request"
+version = "0.0.0"
+edition = "2018"
+
+
+[dependencies]
+tikv-client-common = {path = "../tikv-client-common"}
